refactor: ducks 패턴에 따라 redux 폴더구조 변경
